refactor(web): move agent hooks to lib/agents/hooks.ts, drop AgentRow
- Consolidate useAgents, useAgent, useAdminAgents, usePinnedAgents,
  useCurrentAgent, useAgentController, useIsDefaultAgent,
  useAgentPreferences, useMcpServersForAgentEditor into a single
  src/lib/agents/hooks.ts
- Replace AgentRow (a redundant subset of Persona) with Persona directly
  in AgentsTable and AgentRowActions, removing the toAgentRow mapper
- Add domain/API section splitters to types.ts
