--- web/src/hooks/useIsDefaultAgent.ts
@@ -1,62 +1,0 @@
-"use client";
-
-import { useMemo } from "react";
-import { useSearchParams } from "next/navigation";
-import { SEARCH_PARAM_NAMES } from "@/app/app/services/searchParams";
-import { CombinedSettings } from "@/interfaces/settings";
-import { ChatSession } from "@/app/app/interfaces";
-import { MinimalPersonaSnapshot } from "@/lib/agents/types";
-import { DEFAULT_AGENT_ID } from "@/lib/constants";
-
-/**
- * Determines if the current assistant is the default agent based on:
- * 1. Whether default agent is disabled in settings
- * 2. If URL has an agentId specified
- * 3. Based on the current chat session
- */
-export default function useIsDefaultAgent({
-  liveAgent,
-  existingChatSessionId,
-  selectedChatSession,
-  settings,
-}: {
-  liveAgent: MinimalPersonaSnapshot | undefined;
-  existingChatSessionId: string | null;
-  selectedChatSession: ChatSession | undefined;
-  settings: CombinedSettings | null;
-}) {
-  const searchParams = useSearchParams();
-  const urlAssistantId = searchParams?.get(SEARCH_PARAM_NAMES.PERSONA_ID);
-
-  return useMemo(() => {
-    // If default agent is disabled, it can never be the default agent
-    if (settings?.settings?.disable_default_assistant) {
-      return false;
-    }
-
-    // If URL has an agentId, it's explicitly selected, not default
-    if (
-      urlAssistantId !== null &&
-      urlAssistantId !== DEFAULT_AGENT_ID.toString()
-    ) {
-      return false;
-    }
-
-    // If there's an existing chat session with a persona_id, it's not default
-    if (
-      existingChatSessionId &&
-      selectedChatSession?.persona_id !== DEFAULT_AGENT_ID
-    ) {
-      return false;
-    }
-
-    // If just on `/chat` page, it's the default agent
-    return true;
-  }, [
-    settings?.settings?.disable_default_assistant,
-    urlAssistantId,
-    existingChatSessionId,
-    selectedChatSession?.persona_id,
-    liveAgent?.id,
-  ]);
-}
