+++ web/src/refresh-pages/AgentsNavigationPage.tsx
@@ -4,7 +4,7 @@ import { useMemo, useState, useRef } from "react";
 import AgentCard from "@/sections/agents/AgentCard";
 import { useUser } from "@/providers/UserProvider";
 import { checkUserOwnsAgent } from "@/lib/agents/utils";
-import { useAgents } from "@/hooks/useAgents";
+import { useAgents } from "@/lib/agents/hooks";
 import { MinimalPersonaSnapshot } from "@/lib/agents/types";
 import Text from "@/refresh-components/texts/Text";
 import InputTypeIn from "@/refresh-components/inputs/InputTypeIn";
