+++ web/src/refresh-pages/admin/GroupsPage/SharedGroupResources/index.tsx
@@ -12,7 +12,7 @@ import SimpleCollapsible from "@/refresh-components/SimpleCollapsible";
 import AgentAvatar from "@/refresh-components/avatars/AgentAvatar";
 import { useConnectorStatus } from "@/lib/hooks";
 import { useDocumentSets } from "@/lib/hooks/useDocumentSets";
-import { useAgents } from "@/hooks/useAgents";
+import { useAgents } from "@/lib/agents/hooks";
 import { getSourceMetadata } from "@/lib/sources";
 import type { ValidSources } from "@/lib/types";
 import ResourceContent from "@/refresh-pages/admin/GroupsPage/SharedGroupResources/ResourceContent";
